refactor: Split list
